CR
switch SET_NULL back to DO_NOTHING

We want to keep the repo_id value as-is, even if the associated repo has
disappeared, so as to keep related object grouped together, and to
prevent uniqueness constraint violations.
